LayoutTests/inspector/elements/shadow-dom.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function addShadowRoot()
{
    if (!window.layoutTestController)
        return;
    var element = document.getElementById("delayed-shadow-host");
    var shadowRoot = layoutTestController.ensureShadowRoot(element);
    var shadowChild = document.createElement("div");
    shadowRoot.appendChild(shadowChild);
    var span = document.createElement("span");
    span.textContent = "text";
    shadowChild.appendChild(span);
}

function test()
{
    InspectorTest.runTestSuite([
        function testDumpInitial(next)
        {
            InspectorTest.expandElementsTree(function() {
                InspectorTest.selectNodeWithId("meter", function(node) {
                    InspectorTest.dumpElementsTree(node);
                    next();
                });
            });
        },

        function testDumpUpdated(next) {
            var host;

            function dumpWithShadowRoot()
            {
                InspectorTest.addResult("After shadow root added");
                InspectorTest.runAfterPendingDispatches(function() {
                    InspectorTest.dumpElementsTree(host);
                    next();
                });
            }

            InspectorTest.selectNodeWithId("delayed-shadow-host", function(node) {
                host = node;
                InspectorTest.addResult("Before shadow root added");
                InspectorTest.dumpElementsTree(node);

                WebInspector.domAgent.addEventListener(WebInspector.DOMAgent.Events.ShadowRootUpdated, function() {
                    InspectorTest.expandElementsTree(dumpWithShadowRoot);
                });
                InspectorTest.evaluateInPage("addShadowRoot()");
            });
        }
    ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that shadow DOM tree is properly displayed.
</p>

<meter id="meter" min="0" max="100" value="20"></meter>
<div id="delayed-shadow-host"></div>
</body>
</html>
